Adding a Date&Time subroutine to CFML_GlobalDeps*.f90, a new component to Diffraction_Pattern_Type, and a small correction to CFML_EoS_Mod.f90 for compilation for 64bits.
git-svn-id: https://forge.ill.fr/svn/crysfml@847 06508225-4b59-0410-a2a9-fdab55894d57

+++ Src/CFML_EoS_Mod.f90
@@ -2705,8 +2705,8 @@ Contains
       integer                                       :: iv, inum
       integer, dimension(ncol_data_max)             :: ivet,iorden
 
-      real, dimension(ncol_data_max)                :: vet,vetsd
-      real, dimension(ncol_data_max)                :: rvet
+      real(kind=cp), dimension(ncol_data_max)       :: vet,vetsd
+      real(kind=cp), dimension(ncol_data_max)       :: rvet
 
       !---- Data Files ----!
       character(len=512)                            :: filedat=' '
